Reorganize project into a modular and reusable structure.
This commit refactors the project by organizing all existing files into a new, more structured directory called `modelo-baseline`. This addresses the user's request for a well-organized and modular codebase.

Key changes include:
- Creation of `data`, `src`, and `results` directories to separate data, source code, and model outputs.
- Existing files (`main.py`, `train.py`, `evaluate.py`, `preprocessing.py`, `challenge_data-18-ago.csv`, etc.) have been moved and renamed to fit the new structure.
- All Python scripts have been updated to use relative paths, ensuring that the pipeline remains fully functional after the reorganization.
- A new `predict.py` script has been added to allow for inference on new data.
- The `README.md` has been updated to reflect the new project structure and provide clear instructions for installation and usage.
