--- modelo-baseline/.gitignore
@@ -1,27 +1,0 @@
-# Python
-__pycache__/
-*.pyc
-*.pyo
-*.pyd
-.Python
-env/
-venv/
-
-# Jupyter Notebook
-.ipynb_checkpoints
-
-# Data
-data/
-results/models/
-*.csv
-*.pkl
-*.h5
-
-# OS generated files
-.DS_Store
-.DS_Store?
-._*
-.Spotlight-V100
-.Trashes
-ehthumbs.db
-Thumbs.db
