+++ modelo-baseline/requirements.txt
@@ -1,4 +1,19 @@
-pandas
-scikit-learn
-matplotlib
-numpy
+# Core Data and ML
+pandas==2.3.2
+scikit-learn==1.7.1
+joblib==1.5.1
+pyarrow==14.0.1
+
+# Text Processing
+spacy==3.7.5
+scispacy==0.5.5
+# scispaCy model - installed via URL
+en_core_sci_lg @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz
+
+# Multi-label Learning
+scikit-multilearn==0.2.0
+
+# EDA and Visualization
+matplotlib==3.10.5
+seaborn==0.13.2
+wordcloud==1.9.4
